Getting ready for relaunch
moved files still to be done to their own subdirectory, so can visually
see on the wiki what needs to be done. created placeholder files for all
articles that were on ghost but not on backup. EVERY file needs checking
for wording and images are to be readded using the wiki.js tool.